harjoitukset_ja_tehtavat_pythonilla/tehtava_4/requirements.txt
mysql-connector-python
flask
python-dotenv
psycopg2-binary
requests
pymongo
